+++ src/main/java/com/ohs/monolithic/board/repository/PostViewRepository.java
@@ -4,6 +4,6 @@ import com.ohs.monolithic.board.domain.Board;
 import com.ohs.monolithic.board.domain.PostView;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostViewRepository extends JpaRepository<PostView, Integer> {
-  PostView findByPostIdAndUserId(Integer postId, Long userId);
+public interface PostViewRepository extends JpaRepository<PostView, Long> {
+  PostView findByPostIdAndUserId(Long postId, Long userId);
 }
